+++ sdk/springappdiscovery/arm-springappdiscovery/src/models/mappers.ts
@@ -437,6 +437,13 @@ export const SpringbootserversProperties: coreClient.CompositeMapper = {
           name: "String",
         },
       },
+      labels: {
+        serializedName: "labels",
+        type: {
+          name: "Dictionary",
+          value: { type: { name: "String" } },
+        },
+      },
     },
   },
 };
@@ -710,12 +717,6 @@ export const SpringbootappsProperties: coreClient.CompositeMapper = {
           },
         },
       },
-      siteName: {
-        serializedName: "siteName",
-        type: {
-          name: "String",
-        },
-      },
       springBootVersion: {
         serializedName: "springBootVersion",
         type: {
@@ -758,6 +759,7 @@ export const SpringbootappsProperties: coreClient.CompositeMapper = {
       },
       provisioningState: {
         serializedName: "provisioningState",
+        readOnly: true,
         type: {
           name: "String",
         },
@@ -772,6 +774,13 @@ export const SpringbootappsProperties: coreClient.CompositeMapper = {
               className: "ErrorModel",
             },
           },
+        },
+      },
+      labels: {
+        serializedName: "labels",
+        type: {
+          name: "Dictionary",
+          value: { type: { name: "String" } },
         },
       },
     },
@@ -1159,13 +1168,6 @@ export const SpringbootserversModel: coreClient.CompositeMapper = {
     className: "SpringbootserversModel",
     modelProperties: {
       ...ProxyResource.type.modelProperties,
-      tags: {
-        serializedName: "tags",
-        type: {
-          name: "Dictionary",
-          value: { type: { name: "String" } },
-        },
-      },
       properties: {
         serializedName: "properties",
         type: {
@@ -1183,13 +1185,6 @@ export const SpringbootserversPatch: coreClient.CompositeMapper = {
     className: "SpringbootserversPatch",
     modelProperties: {
       ...ProxyResource.type.modelProperties,
-      tags: {
-        serializedName: "tags",
-        type: {
-          name: "Dictionary",
-          value: { type: { name: "String" } },
-        },
-      },
       properties: {
         serializedName: "properties",
         type: {
@@ -1214,13 +1209,6 @@ export const SpringbootappsModel: coreClient.CompositeMapper = {
           className: "SpringbootappsProperties",
         },
       },
-      tags: {
-        serializedName: "tags",
-        type: {
-          name: "Dictionary",
-          value: { type: { name: "String" } },
-        },
-      },
     },
   },
 };
@@ -1236,13 +1224,6 @@ export const SpringbootappsPatch: coreClient.CompositeMapper = {
         type: {
           name: "Composite",
           className: "SpringbootappsProperties",
-        },
-      },
-      tags: {
-        serializedName: "tags",
-        type: {
-          name: "Dictionary",
-          value: { type: { name: "String" } },
         },
       },
     },
@@ -1358,25 +1339,10 @@ export const SpringbootserversDeleteHeaders: coreClient.CompositeMapper = {
   },
 };
 
-export const SpringbootserversUpdateHeaders: coreClient.CompositeMapper = {
+export const SpringbootappsDeleteHeaders: coreClient.CompositeMapper = {
   type: {
     name: "Composite",
-    className: "SpringbootserversUpdateHeaders",
-    modelProperties: {
-      location: {
-        serializedName: "location",
-        type: {
-          name: "String",
-        },
-      },
-    },
-  },
-};
-
-export const SpringbootappsUpdateHeaders: coreClient.CompositeMapper = {
-  type: {
-    name: "Composite",
-    className: "SpringbootappsUpdateHeaders",
+    className: "SpringbootappsDeleteHeaders",
     modelProperties: {
       location: {
         serializedName: "location",