+++ sdk/springappdiscovery/arm-springappdiscovery/CHANGELOG.md
@@ -1,15 +1,40 @@
 # Release History
+    
+## 1.0.0-beta.2 (2024-03-25)
+    
+**Features**
 
-## 1.0.0-beta.2 (Unreleased)
+  - Added operation Springbootapps.beginCreateOrUpdate
+  - Added operation Springbootapps.beginCreateOrUpdateAndWait
+  - Added operation Springbootapps.beginDelete
+  - Added operation Springbootapps.beginDeleteAndWait
+  - Added operation Springbootapps.update
+  - Added operation Springbootservers.update
+  - Added Interface SpringbootappsCreateOrUpdateOptionalParams
+  - Added Interface SpringbootappsDeleteHeaders
+  - Added Interface SpringbootappsDeleteOptionalParams
+  - Added Type Alias SpringbootappsCreateOrUpdateResponse
+  - Added Type Alias SpringbootappsDeleteResponse
+  - Interface SpringbootappsProperties has a new optional parameter labels
+  - Interface SpringbootserversProperties has a new optional parameter labels
 
-### Features Added
+**Breaking Changes**
 
-### Breaking Changes
-
-### Bugs Fixed
-
-### Other Changes
-
+  - Removed operation Springbootapps.beginUpdate
+  - Removed operation Springbootapps.beginUpdateAndWait
+  - Removed operation Springbootservers.beginUpdate
+  - Removed operation Springbootservers.beginUpdateAndWait
+  - Interface SpringbootappsModel no longer has parameter tags
+  - Interface SpringbootappsPatch no longer has parameter tags
+  - Interface SpringbootappsProperties no longer has parameter siteName
+  - Interface SpringbootappsUpdateOptionalParams no longer has parameter resumeFrom
+  - Interface SpringbootappsUpdateOptionalParams no longer has parameter updateIntervalInMs
+  - Interface SpringbootserversModel no longer has parameter tags
+  - Interface SpringbootserversPatch no longer has parameter tags
+  - Interface SpringbootserversUpdateOptionalParams no longer has parameter resumeFrom
+  - Interface SpringbootserversUpdateOptionalParams no longer has parameter updateIntervalInMs
+    
+    
 ## 1.0.0-beta.1 (2024-02-19)
 
 The package of @azure/arm-springappdiscovery is using our next generation design principles. To learn more, please refer to our documentation [Quick Start](https://aka.ms/azsdk/js/mgmt/quickstart).